moved excel.py to utils, created init and updated references in import_service_v2.py
